playground: let clients request vet check in same HTTP request as compile+run
Also, move the tests to their own file and extend them a bit, give
them names, and make the test step more verbose (only visible in
docker build anyway). They were hogging up the sandbox file.

Updates golang/go#31970

Reviewed-on: https://go-review.googlesource.com/c/playground/+/176598
