+++ go.mod
@@ -5,5 +5,6 @@ go 1.12
 require (
 	cloud.google.com/go v0.38.0
 	github.com/bradfitz/gomemcache v0.0.0-20190329173943-551aad21a668
-	golang.org/x/tools v0.0.0-20190509153222-73554e0f7805
+	golang.org/x/net v0.0.0-20190509222800-a4d6f7feada5 // indirect
+	golang.org/x/tools v0.0.0-20190513214131-2a413a02cc73
 )
